refactor: update type aliases and usage recording logic
*   Changed type aliases in `core/grant_manager.go` and `core/types.go` to use the direct types from the `models` package instead of pointers to them
*   Updated `SharedWith` field values in `allowance.go` from 0 to 1 for cardinality semantics
*   Modified `updateDailyUsage` function in `base.go` to accept `int64` bytes for proper handling of storage removals
*   Implemented proper storage delta calculations and zero clamping in `updateDailyUsage`
*   Added `getUsageByType` helper function in `base.go` to aggregate usage across all time
*   Refactored `getCurrentUsage` in `base.go` to calculate total usage instead of daily usage
*   Added `getAggregatedUsageByType` function in `hard_limits.go` to support total usage limits
*   Updated usage recording in `threshold.go` to use signed delta for storage changes

+++ internal/service/policies/unlimited.go
@@ -76,7 +76,7 @@ func (u *UnlimitedPolicyEnforcer) RecordUpload(userID, uploadID uint, bytes uint
 	}
 
 	// Update daily aggregated usage
-	return u.updateDailyUsage(userID, models.UsageTypeUpload, bytes)
+	return u.updateDailyUsage(userID, models.UsageTypeUpload, int64(bytes))
 }
 
 // RecordDownload simply records usage without any limit checking
@@ -104,7 +104,7 @@ func (u *UnlimitedPolicyEnforcer) RecordDownload(userID, uploadID uint, bytes ui
 	}
 
 	// Update daily aggregated usage
-	return u.updateDailyUsage(userID, models.UsageTypeDownload, bytes)
+	return u.updateDailyUsage(userID, models.UsageTypeDownload, int64(bytes))
 }
 
 // RecordStorageChange simply records usage without any limit checking
@@ -142,7 +142,7 @@ func (u *UnlimitedPolicyEnforcer) RecordStorageChange(userID, uploadID uint, byt
 	}
 
 	// Update daily aggregated usage
-	return u.updateDailyUsage(userID, usageType, recordBytes)
+	return u.updateDailyUsage(userID, usageType, int64(recordBytes))
 }
 
 // GetDetailedUsage delegates to the base enforcer
